requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
faiss-cpu==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
pypdf2==3.0.1
pypdf==3.17.0
tiktoken==0.5.2
numpy==1.24.3
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.0
aiofiles==23.2.1
